requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pandas>=1.2.4
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
Pillow>=8.2.0
opencv-python>=4.5.2
tqdm>=4.61.0
PyYAML>=5.4.1 